improve pykwasm scripts
