+++ pykwasm/pyproject.toml
@@ -12,6 +12,7 @@ authors = [
 
 [tool.poetry.scripts]
 wasm = "pykwasm.run_wasm:main"
+wasm2kore = "pykwasm.wasm2kore:main"
 wasm2kast = "pykwasm.wasm2kast:main"
 kwasm = "pykwasm.scripts.kwasm:main"
 kwasm-convert = "pykwasm.scripts.convert:main"
